requirements.txt
langchain-groq
# The langchain-groq library is used to integrate Grok-hosted large language models into a LangChain-based application. It acts as a bridge between LangChain and the Grok Cloud API, allowing the project to send prompts to Grok models such as LLaMA and receive generated responses. In this project, langchain-groq is responsible for connecting the AI agents to the selected Grok LLM, enabling natural language understanding and text generation as the core intelligence of the system.


langchain-community
# The langchain-community package provides a collection of community-maintained integrations, tools, and utilities that extend the functionality of LangChain. It includes connectors for external services such as web search tools, document loaders, and retrievers that are not part of the LangChain core. In this project, it is typically used to integrate third-party tools like online search services (for example, Tavily) so that the AI agents can fetch real-time information from the internet and enrich their responses beyond the LLM’s training data.

python-dotenv


uvicorn
# uvicorn is an ASGI (Asynchronous Server Gateway Interface) server used to run FastAPI applications. It provides a high-performance, asynchronous web server capable of handling multiple concurrent requests efficiently. In this project, uvicorn is responsible for serving the FastAPI backend, listening for incoming HTTP requests from the Streamlit frontend, and delivering responses generated by the AI agents.

fastapi

pydantic
# pydantic is a data validation and data modeling library that is tightly integrated with FastAPI. It is used to define request and response schemas using Python classes, ensuring that incoming user data is correctly structured and type-safe. In this project, pydantic models are used to validate inputs such as the agent type and user query, preventing invalid data from reaching the core application logic and improving overall reliability.

streamlit 

langgraph
# langgraph is an agentic AI framework built on top of LangChain. While LangChain focuses on interacting with language models and tools, LangGraph enables the creation of structured AI agents and workflows using graph-based execution. In this project, langgraph is used to define the behavior, decision flow, and coordination of AI agents, making it possible to build multi-agent systems where each agent follows a controlled sequence of reasoning and actions.

langchain_core